requirements.txt
Adafruit-DHT==1.4.0
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.7
colorzero==2.0
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==4.2.8
exceptiongroup==1.1.3
executing==2.0.1
fonttools==4.44.0
googletrans==4.0.0rc1
gpiozero==2.0
gTTS==2.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
jupyter-client==8.5.0
jupyter-core==5.4.0
kiwisolver==1.4.5
matplotlib==3.8.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.13
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-vlc==3.0.18122
pyttsx3==2.90
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
RPi.GPIO==0.7.1
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
spidev==3.6
sqlparse==0.4.4
stack-data==0.6.3
textblob==0.17.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
typing-extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
zipp==3.17.0
